Harden models against real Presscart API data
Live integration testing revealed several mismatches between the
documented response shapes and what the API actually returns. This
commit loosens the models where the real data demanded it:

Money fields (int → float):
- Order.total, subtotal, processing_fee, discount, credits_applied
- LineItem.price
- OrderItem.commission_amount, sale_price, product_price,
  vendor_cost, internal_cost, reseller_price
- CheckoutRequest.discount
- Price.unit_amount
- Campaign.budget
Order endpoints return amounts in dollars (e.g. 154.5), not cents.
Product endpoints still return cents. Units vary by endpoint — we
now surface whatever the API sends, and document the difference.

Nullable enums:
- OutletChannelSummary.channel_type / placement_type
- OutletChannel.channel_type / placement_type
- IncludeItem.channel_type / placement_type
The API occasionally returns null for these on channels where the
placement isn't yet pinned down.

Lenient parsing:
- PresscartModel gains a before-validator that coerces empty strings
  to None. Real responses surface "" in place of null for optional
  booleans/dates (e.g. OutletChannel.do_follow_links_allowed="").

Filter encoding:
- serialize_filters now emits list-valued filters using PHP/Rails
  bracket notation: filters[channel_types][]=WEBSITE. Scalar filters
  are unchanged. Caught by the API rejecting list filters as
  "Expected array, received string".

Docs updated in lockstep: Models Reference now warns about mixed
monetary units, notes nullable IncludeItem enums, and documents the
empty-string coercion. Orders resource doc carries a note about
dollar-unit totals and corrects CheckoutRequest.discount type.

All 47 unit tests + lint + type-check + docs build pass.

+++ src/pypresscart/models/campaigns.py
@@ -50,7 +50,7 @@ class Campaign(PresscartModel):
     status: str | None = None
     goals: str | None = None
     target_date: datetime | None = None
-    budget: int | None = None
+    budget: float | None = None
     created_at: datetime | None = None
     updated_at: datetime | None = None
     deleted_at: datetime | None = None
